ml/requirements.txt
ultralytics
fastapi
uvicorn
python-multipart
numpy
pillow
opencv-python-headless

